requirements.txt
aiogram
python-dotenv
qrcode
asyncpg
pillow
pandas
openpyxl